issue-174: Code refactoring
